requirements.txt
Flask
PyPDF2
python-docx
scikit-learn
nltk
gunicorn
